avfrontend/backend/requirements.txt
langchain-core==1.2.1
langchain-unstructured==1.0.0
langchain-chroma==1.1.0
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
chromadb==1.4.0
sentence-transformers==5.2.0
transformers==4.57.3
torch==2.9.1
python-dotenv==1.2.1
unstructured-client==0.42.6
aiofiles==25.1.0
pydantic==2.12.5
